requirements.txt
SpeechRecognition>=3.10.0
pyttsx3>=2.90
wikipedia>=1.4.0
requests>=2.31.0
wolframalpha>=5.0.0
openai>=1.0.0
PyYAML>=6.0
llama-cpp-python>=0.2.0; platform_system != "Windows"
